Move test and union equivalence
